core: moved safeFunctionName creation under Misc
