+++ packages/restlessness-core/src/Misc/index.ts
@@ -1,5 +1,6 @@
 import { existsSync, promises as fs } from 'fs';
 import path from 'path';
+import crypto from 'crypto';
 
 export default class Misc {
   static capitalize = (s: string): string => `${s[0].toUpperCase()}${s.slice(1)}`;
@@ -35,5 +36,26 @@ export default class Misc {
         }
       }
     }
+  };
+  static createAwsSafeFunctionName = (functionId: string, fullServiceName: string) => {
+    const id = functionId; //@TODO check for symbols presence (-./ etc..)
+    /**
+     * The 4 xxxx stand for "dev" or "prod", based on which stage deployment will be selected
+     * We use 4 x for worst case scenario, that is "prod", since we need to check this string length
+     * and to avoid it will reach 64 chars, since AWS complains about that
+     */
+    let safeFunctionName: string = id;
+    const awsLambdaName: string = `${id}LambdaFunction`;
+    if (awsLambdaName.length > 63) {
+      const hash = crypto.createHash('md5').update(id).digest('hex');
+      safeFunctionName = `${id.substring(0, 21)}${hash.substring(0, 3)}${id.substring(id.length - 21)}`;
+    }
+    const awsFunctionName: string = `${fullServiceName}-xxxx-${id}`;
+    if (awsFunctionName.length > 63) {
+      const hash = crypto.createHash('md5').update(id).digest('hex');
+      const chars = Math.floor((64 - `${fullServiceName}-xxxx-`.length - 10) / 2);
+      safeFunctionName = `${id.substring(0, chars)}${hash.substring(0, 3)}${id.substring(id.length - chars)}`;
+    }
+    return safeFunctionName;
   };
 }
